+++ src/push/mod.rs
@@ -16,7 +16,10 @@ pub mod varint;
 pub use callbacks::*;
 pub use push_client::*;
 pub use push_message::*;
-pub use tick_util::{convert_trade_tick, get_part_name_by_code, get_part_short_name_by_code, get_trade_cond_by_code, is_us_stock_symbol, PushTick, PushTradeTick};
+pub use tick_util::{
+    convert_trade_tick, get_part_name_by_code, get_part_short_name_by_code, get_trade_cond_by_code,
+    is_us_stock_symbol, PushTick, PushTradeTick,
+};
 
 #[cfg(test)]
 mod tests;